feat: Add support for outer.control in DITA
Previously, DITA files using something like @conref to a parent directory lead to an error as outer.control was not set in the internal call of the DITA command. This is now supported. Some additional code changes were required to find the generated .html file as this means that the .html file can now be in a subdirectory of the temporary folder.