(GH-597) Find config in home folder as last resort
Prior to this change, the [documentation][01] noted that when no closer
configuration can be found, Vale will default to using the configuration
file in the user's home directory. However, the implementation for the
`loadConfig()` function only recursed through the current working
directory's ancestors.

This meant that if the working directory didn't have the user's home
directory as an ancestor, no configuration would be found and the
user would get an error even when they had defined a configuration file
in their home directory.

This change:

- Updates the `loadConfig()` function to explicitly search the user's
  home directory if no configuration file is discovered in an ancestor
  folder of the current working directory.
- Maintains all other prior functionality.
- Resolves #597.

[01]: https://vale.sh/docs/topics/config/#search-process