feat: Allow users to override comment delimiters
Allow users to specify the comment delimiters they are using in their
documentation. vale replaces these with HTML comment tags before
linting, making it possible to control style rules for specific passages
of prose in file formats that use non-HTML comment syntax. This is
critical for controlling style rules within a page in MDX, and
potentially other formats as well.

This example configures the `CommentDelimiters` field for `*.md` files,
indicating that `{/*` and `*/}` are the custom comment delimiters:

```ini
[*.md]
CommentDelimiters = "{/*,*/}"
```

Internally, custom delimiters are represented as a `[2]string`, and it
is only possible to configure one set of custom comment delimiters for a
given format block.

More specific changes:
- Refactor `applyPatterns`. Remove the method receiver and take only the
  necessary fields of `*core.Config` as parameters. This makes it easier
  to test `applyPatterns` without mocking an entire `*core.Config`. Also
  extract functions for `applyInlinePatterns` and `applyBlockPatterns`
  so we can use fewer arguments in a single function.
- Add `applyCommentPatterns`, which works similarly to
  `applyInlinePatterns` and `applyBlockPatterns`, but for substituting
  comments.
- For tests, import `https//github.com/stretchr/testify/assert`, a
  popular testing library, to get richer test output.

Closes #762